+++ internal/paw/env.go
@@ -1,0 +1,6 @@
+package paw
+
+const (
+	ENV_HOME    = "PAW_HOME"    // The env var name can be used to override the Paw HOME directory
+	ENV_SESSION = "PAW_SESSION" // The env var name can be used to specify a Paw session ID
+)