+++ internal/ui/items.go
@@ -10,7 +10,9 @@ import (
 	"strings"
 
 	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/canvas"
 	"fyne.io/fyne/v2/container"
+	"fyne.io/fyne/v2/theme"
 	"fyne.io/fyne/v2/widget"
 
 	"lucor.dev/paw/internal/icon"
@@ -76,12 +78,25 @@ func (iw *itemsWidget) makeList(selectedItem paw.Item, opts *paw.VaultFilterOpti
 			return len(itemMetadata)
 		},
 		func() fyne.CanvasObject {
-			return container.NewHBox(widget.NewIcon(icon.LockOutlinedIconThemed), widget.NewLabel("Identity label"))
+			icon := widget.NewIcon(icon.LockOutlinedIconThemed)
+			return newItemListContainer(icon, &canvas.Text{
+				Text:      "Item name",
+				TextSize:  theme.TextSize(),
+				TextStyle: fyne.TextStyle{Bold: true},
+			}, &canvas.Text{
+				Text:     "Item subtitle",
+				TextSize: theme.CaptionTextSize(),
+			})
 		},
 		func(id int, obj fyne.CanvasObject) {
 			metadata := &Metadata{Metadata: itemMetadata[id]}
 			obj.(*fyne.Container).Objects[0].(*widget.Icon).SetResource(metadata.Icon())
-			obj.(*fyne.Container).Objects[1].(*widget.Label).SetText(metadata.String())
+			title := obj.(*fyne.Container).Objects[1].(*canvas.Text)
+			title.Text = metadata.String()
+			title.Refresh()
+			subtitle := obj.(*fyne.Container).Objects[2].(*canvas.Text)
+			subtitle.Text = metadata.Subtitle
+			subtitle.Refresh()
 		})
 
 	list.OnSelected = func(id widget.ListItemID) {
@@ -90,4 +105,46 @@ func (iw *itemsWidget) makeList(selectedItem paw.Item, opts *paw.VaultFilterOpti
 	}
 
 	return list
+}
+
+var (
+	iconBoxSize = fyne.NewSize(36, 36)
+	iconSize    = fyne.NewSize(24, 24)
+)
+
+// itemListLayout is a custom layout for the item list
+type itemListLayout struct {
+	icon     *widget.Icon
+	title    *canvas.Text
+	subtitle *canvas.Text
+}
+
+// MinSize returns the minimum size required by the custom layout
+func (l *itemListLayout) MinSize(objects []fyne.CanvasObject) fyne.Size {
+	boxSizeMaxHeight := fyne.Max(l.title.MinSize().Height, l.subtitle.MinSize().Height)
+	boxSizeMaxWidth := fyne.Max(l.title.MinSize().Width, l.subtitle.MinSize().Width)
+	return fyne.NewSize(iconBoxSize.Width+boxSizeMaxWidth, fyne.Max(iconBoxSize.Height, boxSizeMaxHeight*2))
+}
+
+// Layout arranges the objects in the custom layout
+func (l *itemListLayout) Layout(objects []fyne.CanvasObject, containerSize fyne.Size) {
+	l.icon.Resize(iconSize)
+	l.icon.Move(fyne.NewPos((iconBoxSize.Width-l.icon.MinSize().Width)/2, (containerSize.Height-l.icon.MinSize().Height)/2))
+	boxLeft := iconBoxSize.Width + theme.InnerPadding()
+	if l.subtitle.Text == "" {
+		l.title.Move(fyne.NewPos(boxLeft, (containerSize.Height-l.title.MinSize().Height)/2))
+		return
+	}
+	l.title.Move(fyne.NewPos(boxLeft, theme.SeparatorThicknessSize()))
+	l.subtitle.Move(fyne.NewPos(boxLeft, containerSize.Height/2))
+}
+
+// newItemListContainer returns a new container for the item list
+func newItemListContainer(icon *widget.Icon, title, subtitle *canvas.Text) *fyne.Container {
+	layout := &itemListLayout{
+		icon:     icon,
+		title:    title,
+		subtitle: subtitle,
+	}
+	return container.New(layout, icon, title, subtitle)
 }